- Tweaked fmpz_mat_hnf_minors a bit (some shortcuts when reducing
  rows). This brings the example from Frank from 3.3s to 1.6s on my
  machine.
- Added fmpz_mat_hnf_minors_transform.
- As fmpz_mat_hnf_minors_transform is usually much faster as the
  previous fmpz_mat_transform, I added some logic in fmpz_mat_transform
  to call it when applicable. Since it involves a rank computation mod p
  as a precomputation, this makes the function a bit slower in case the
  matrix has not full rank, but otherwise speeds up fmpz_mat_transform
  tremendously (In Frank's example from 120s to 6s on my machine).