@wbhart 2017-10-13
The only potential issue I see is the very high iteration count in the test code. How long does this test take to run now?