@thofma 2017-10-13
I did not change the number of iterations. It was 10000 before and now it is 5000 and 5000.