@wbhart 2017-10-13
Thanks. And bravo on the new code!